requirements.txt
python-telegram-bot==21.6
openai==1.51.0
python-dotenv==1.0.1
httpx==0.27.2
fastapi==0.115.0
uvicorn==0.31.0
docker==7.1.0
